changed parseInt to parseFloat

+++ js/challenge-3.js
@@ -10,7 +10,7 @@ const divider = '----------------------------------'
 
 // Only change below this line
 
-const owed = "R " + (parseInt(leoBalance * -1) + parseInt(sarahBalance * -1)).toFixed(2) //added * -1 to create +
+const owed = "R " + (parseFloat(leoBalance * -1) + parseFloat(sarahBalance * -1)).toFixed(2) //added * -1 to create +
 const leo = `${leoName} ${leoSurname} (Owed: R  ${(leoBalance * -1).toFixed(2)})`
 const sarah =`${sarahName} ${sarahSurname} (Owed: R  ${(sarahBalance * -1).toFixed(2)})` //gives decimal
 const total = "Total amount owed: "
